+++ neo/idlib/math/Math.h
@@ -186,8 +186,8 @@ public:
 	static float				Rint( float f );			// returns the nearest integer
 	static int					Ftoi( float f );			// float to int conversion
 	static int					FtoiFast( float f );		// fast float to int conversion but uses current FPU round mode (default round nearest)
-	static unsigned long		Ftol( float f );			// float to long conversion
-	static unsigned long		FtolFast( float );			// fast float to long conversion but uses current FPU round mode (default round nearest)
+	static unsigned int		Ftol( float f );			// float to long conversion
+	static unsigned int		FtolFast( float );			// fast float to long conversion but uses current FPU round mode (default round nearest)
 
 	static signed char			ClampChar( int i );
 	static signed short			ClampShort( int i );
@@ -241,11 +241,11 @@ private:
 
 ID_INLINE float idMath::RSqrt( float x ) {
 
-	long i;
+	int i;
 	float y, r;
 
 	y = x * 0.5f;
-	i = *reinterpret_cast<long *>( &x );
+	i = *reinterpret_cast<int *>( &x );
 	i = 0x5f3759df - ( i >> 1 );
 	r = *reinterpret_cast<float *>( &i );
 	r = r * ( 1.5f - r * r * y );
@@ -827,14 +827,14 @@ ID_INLINE int idMath::FtoiFast( float f ) {
 #endif
 }
 
-ID_INLINE unsigned long idMath::Ftol( float f ) {
-	return (unsigned long) f;
+ID_INLINE unsigned int idMath::Ftol( float f ) {
+	return (unsigned int) f;
 }
 
-ID_INLINE unsigned long idMath::FtolFast( float f ) {
+ID_INLINE unsigned int idMath::FtolFast( float f ) {
 #ifdef _WIN32
 	// FIXME: this overflows on 31bits still .. same as FtoiFast
-	unsigned long i;
+	unsigned int i;
 	__asm fld		f
 	__asm fistp		i		// use default rouding mode (round nearest)
 	return i;
@@ -857,7 +857,7 @@ ID_INLINE unsigned long idMath::FtolFast( float f ) {
 						  : "m" (f) );
 	return i;
 #else
-	return (unsigned long) f;
+	return (unsigned int) f;
 #endif
 }
