+++ neo/sys/sys_public.h
@@ -49,8 +49,7 @@ If you have questions concerning this license or the applicable additional terms
 #define ALIGN16( x )					__declspec(align(16)) x
 #define PACKED
 
-#define _alloca16( x )					((void *)((((int)_alloca( (x)+15 )) + 15) & ~15))
-
+#define _alloca16( x )          ((void *)((((uintptr_t)_alloca( (x)+15 )) + 15) & ~15))
 #define PATHSEPERATOR_STR				"\\"
 #define PATHSEPERATOR_CHAR				'\\'
 
@@ -84,8 +83,7 @@ If you have questions concerning this license or the applicable additional terms
 #endif
 
 #define _alloca							alloca
-#define _alloca16( x )					((void *)((((size_t)alloca( (x)+15 )) + 15) & ~15))
-
+#define _alloca16( x )          ((void *)((((uintptr_t)alloca( (x)+15 )) + 15) & ~15))
 #define PATHSEPERATOR_STR				"/"
 #define PATHSEPERATOR_CHAR				'/'
 
@@ -120,8 +118,7 @@ If you have questions concerning this license or the applicable additional terms
 #endif
 
 #define _alloca							alloca
-#define _alloca16( x )					((void *)((((size_t)alloca( (x)+15 )) + 15) & ~15))
-
+#define _alloca16( x )          ((void *)((((uintptr_t)alloca( (x)+15 )) + 15) & ~15))
 #define ALIGN16( x )					x
 #define PACKED							__attribute__((packed))
 
@@ -331,9 +328,9 @@ const char *	Sys_GetCallStackCurAddressStr( int depth );
 void			Sys_ShutdownSymbols( void );
 
 // DLL loading, the path should be a fully qualified OS path to the DLL file to be loaded
-size_t			Sys_DLL_Load( const char *dllName );
-void *			Sys_DLL_GetProcAddress( size_t dllHandle, const char *procName );
-void			Sys_DLL_Unload( size_t dllHandle );
+uintptr_t			Sys_DLL_Load( const char *dllName );
+void *			Sys_DLL_GetProcAddress( uintptr_t dllHandle, const char *procName );
+void			Sys_DLL_Unload( uintptr_t dllHandle );
 
 // event generation
 void			Sys_GenerateEvents( void );
@@ -563,9 +560,9 @@ public:
 	virtual const char *	GetCallStackCurStr( int depth ) = 0;
 	virtual void			ShutdownSymbols( void ) = 0;
 
-	virtual int				DLL_Load( const char *dllName ) = 0;
-	virtual void *			DLL_GetProcAddress( int dllHandle, const char *procName ) = 0;
-	virtual void			DLL_Unload( int dllHandle ) = 0;
+	virtual uintptr_t		DLL_Load( const char *dllName ) = 0;
+	virtual void *			DLL_GetProcAddress( uintptr_t dllHandle, const char *procName ) = 0;
+	virtual void			DLL_Unload( uintptr_t dllHandle ) = 0;
 	virtual void			DLL_GetFileName( const char *baseName, char *dllName, int maxLength ) = 0;
 
 	virtual sysEvent_t		GenerateMouseButtonEvent( int button, bool down ) = 0;
